카드 UI에 대한 부분을 재사용 컴포넌트로 분리
