+++ src/components/Card.tsx
@@ -1,0 +1,62 @@
+import type CardNumberType from '../types/CardNumberType';
+import type ExpirationDateType from '../types/ExpirationDateType';
+
+import {
+	FIRST_NUMBER, FOURTH_NUMBER, SECOND_NUMBER, THIRD_NUMBER,
+} from '../constants/cardNumber';
+import {MONTH, YEAR} from '../constants/expirationDate';
+import {OWNER_NAME_DEFAULT} from '../constants/default';
+import {CARD_NUMBER_LIMIT, EXPIRATION_DATE_LIMIT} from '../constants/limit';
+import {HYPHEN, SLASH} from '../constants/specialCharacter';
+import {EXPIRATION_DATE_MONTH_PLACEHOLDER, EXPIRATION_DATE_YEAR_PLACEHOLDER} from '../constants/placeHolder';
+
+type CardProps = {
+	cardNumber: CardNumberType;
+	ownerName: string;
+	expirationDate: ExpirationDateType;
+};
+
+export default function Card({cardNumber, ownerName, expirationDate}: CardProps) {
+	const specialCharacter = (
+		value: string,
+		limit: number,
+		specialCharacter: string,
+	) => value.length === limit ? specialCharacter : '';
+
+	return (
+		<>
+			<div className='card-top'>
+				<span className='card-text'>클린카드</span>
+			</div>
+			<div className='card-middle'>
+				<div className='small-card__chip'></div>
+			</div>
+			<div className='card-bottom'>
+				<div className='card-bottom__number'>
+					<span className='card-text'>
+						{cardNumber[FIRST_NUMBER]
+						+ specialCharacter(cardNumber[FIRST_NUMBER], CARD_NUMBER_LIMIT, HYPHEN)}
+						{cardNumber[SECOND_NUMBER]
+						+ specialCharacter(cardNumber[SECOND_NUMBER], CARD_NUMBER_LIMIT, HYPHEN)}
+						{cardNumber[THIRD_NUMBER]
+						+ specialCharacter(cardNumber[THIRD_NUMBER], CARD_NUMBER_LIMIT, HYPHEN)}
+						{cardNumber[FOURTH_NUMBER]}
+					</span>
+
+				</div>
+				<div className='card-bottom__info'>
+					<span className='card-text'>{ownerName || OWNER_NAME_DEFAULT}</span>
+					<span className='card-text'>
+						{
+							expirationDate[MONTH] || expirationDate[YEAR]
+								? `${expirationDate[MONTH]} 
+								   ${specialCharacter(expirationDate[MONTH], EXPIRATION_DATE_LIMIT, SLASH)}
+								   ${expirationDate[YEAR]}`
+								: `${EXPIRATION_DATE_MONTH_PLACEHOLDER} ${SLASH} ${EXPIRATION_DATE_YEAR_PLACEHOLDER}`
+						}
+					</span>
+				</div>
+			</div>
+		</>
+	);
+}
