+++ src/types/ExpirationDateType.ts
@@ -1,0 +1,8 @@
+import {type MONTH, type YEAR} from '../constants/expirationDate';
+
+type ExpirationDateType = {
+	[MONTH]: string;
+	[YEAR]: string;
+};
+
+export default ExpirationDateType;
